+++ hat/test/org/hatdex/hat/api/HATTestContext.scala
@@ -40,6 +40,7 @@ import org.hatdex.hat.api.service._
 import org.hatdex.hat.api.service.applications.{ TestApplicationProvider, TrustedApplicationProvider }
 import org.hatdex.hat.authentication.HatApiAuthEnvironment
 import org.hatdex.hat.authentication.models.HatUser
+import org.hatdex.hat.clients._
 import org.hatdex.hat.dal.HatDbSchemaMigration
 import org.hatdex.hat.phata.models.MailTokenUser
 import org.hatdex.hat.resourceManagement.{ FakeHatConfiguration, FakeHatServerProvider, HatServer, HatServerProvider }
@@ -233,6 +234,8 @@ trait HATTestContext extends PostgresqlSpec with MockitoSugar with BeforeAndAfte
       bind[AsyncCacheApi].annotatedWith(new NamedCacheImpl("user-cache")).to[FakeCache]
       bind[AsyncCacheApi].to[FakeCache]
       bind[LoggingProvider].toInstance(new MockLoggingProvider(mockLogger))
+      bind[AuthServiceClient].to[AuthServiceWsClient]
+      bind[AdjudicatorClient].to[AdjudicatorWsClient]
     }
 
     @Provides
